+++ Rubrik/Private/Submit-Request.ps1
@@ -8,8 +8,35 @@
     if ($PSCmdlet.ShouldProcess($id, $resources.Description)) {
         try {
             Write-Verbose -Message 'Submitting the request'
-            # Because some calls require more than the default payload limit of 2MB, ExpandPayload dynamically adjusts the payload limit
-            $result = ExpandPayload -response (Invoke-RubrikWebRequest -Uri $uri -Headers $header -Method $method -Body $body)
+            if ($resources.method -eq 'Delete') {
+                # Delete operations generally have no response body, skip JSON formatting and store the response from Invoke-WebRequest
+                $response = Invoke-RubrikWebRequest -Uri $uri -Headers $header -Method $method -Body $body
+                # Parse response to verify there is nothing in the body
+                $result = ExpandPayload -response $response
+                # If $result is null, build a $result object to return to the user. Otherwise, $result will be returned.
+                if($null -eq $result) {   
+                    # If if HTTP status code matches our expected result, build a PSObject reflecting success
+                    if($response.StatusCode -eq [int]$resources.Success) {
+                        $result = @{
+                            Status = 'Success'
+                            HTTPStatusCode = $response.StatusCode
+                        }
+                    }
+                    # If a different HTTP status is returned, surface that information to the user
+                    # This code may never run due to non-200 HTTP codes throwing an HttpResponseException
+                    else {
+                        $result = @{
+                            Status = 'Error'
+                            HTTPStatusCode = $response.StatusCode
+                            HTTPStatusDescription = $response.StatusCode
+                        }
+                    }
+                }
+            }
+            else {
+                # Because some calls require more than the default payload limit of 2MB, ExpandPayload dynamically adjusts the payload limit
+                $result = ExpandPayload -response (Invoke-RubrikWebRequest -Uri $uri -Headers $header -Method $method -Body $body)
+            }
         }
         catch {
             switch -Wildcard ($_) {